refactoring mvc
