+++ JS/index.js
@@ -1,0 +1,7 @@
+import {homepageController} from "./controller/homepageController.js";
+
+fetch("../json/data.json")
+    .then((response) => response.json())
+    .then((data) => {
+        homepageController.init(data);
+    });
